+++ src/upper_camel.rs
@@ -1,38 +1,38 @@
 use crate::{capitalize, transform};
 
-/// This trait defines a camel case conversion.
+/// This trait defines an upper camel case conversion.
 ///
-/// In CamelCase, word boundaries are indicated by capital letters, including
-/// the first word.
+/// In UpperCamelCase, word boundaries are indicated by capital letters,
+/// including the first word.
 ///
 /// ## Example:
 ///
 /// ```rust
-/// use heck::ToCamelCase;
+/// use heck::ToUpperCamelCase;
 ///
 /// let sentence = "We are not in the least afraid of ruins.";
-/// assert_eq!(sentence.to_camel_case(), "WeAreNotInTheLeastAfraidOfRuins");
+/// assert_eq!(sentence.to_upper_camel_case(), "WeAreNotInTheLeastAfraidOfRuins");
 /// ```
-pub trait ToCamelCase: ToOwned {
-    /// Convert this type to camel case.
-    fn to_camel_case(&self) -> Self::Owned;
+pub trait ToUpperCamelCase: ToOwned {
+    /// Convert this type to upper camel case.
+    fn to_upper_camel_case(&self) -> Self::Owned;
 }
 
-impl ToCamelCase for str {
-    fn to_camel_case(&self) -> String {
+impl ToUpperCamelCase for str {
+    fn to_upper_camel_case(&self) -> String {
         transform(self, capitalize, |_| {})
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::ToCamelCase;
+    use super::ToUpperCamelCase;
 
     macro_rules! t {
         ($t:ident : $s1:expr => $s2:expr) => {
             #[test]
             fn $t() {
-                assert_eq!($s1.to_camel_case(), $s2)
+                assert_eq!($s1.to_upper_camel_case(), $s2)
             }
         };
     }